+++ Lesson24Web/src/main/java/Actions.java
@@ -1,3 +1,3 @@
 public enum Actions {
-    SUM, SUBTRACION, MULTIPLY, DIVISION,
+    SUM, SUBTRACTION, MULTIPLY, DIVISION;
 }
