--- Lesson24Web/src/main/java/Calculator.java
@@ -1,27 +1,0 @@
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class Calculator {
-    private int a;
-    private int b;
-
-    public int sum(int a, int b) {
-        return a + b;
-    }
-
-    public int subtraction(int a, int b) {
-        return a - b;
-    }
-
-    public int multiply(int a, int b) {
-        return a * b;
-    }
-
-    public int division(int a, int b) {
-        return a / b;
-    }
-}
