+++ Lesson24Web/src/main/java/HelloServlet.java
@@ -1,7 +1,9 @@
 import java.io.*;
 
+import jakarta.servlet.ServletException;
 import jakarta.servlet.http.*;
 import jakarta.servlet.annotation.*;
+import model.Calculator;
 
 @WebServlet(name = "HelloServlet", value = "/my-servlet2023")
 public class HelloServlet extends HttpServlet {
@@ -13,12 +15,19 @@ public class HelloServlet extends HttpServlet {
 
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
         response.setContentType("text/html");
-
-        // Hello
         PrintWriter out = response.getWriter();
         out.println("<html><body>");
         out.println("<h1>" + message + "</h1>");
         out.println("</body></html>");
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        Calculator calculator = new Calculator();
+        calculator.sum(2, 5);
+        calculator.subtraction(10, 5);
+        calculator.division(100, 101);
+        calculator.multiply(10, 50);
     }
 
     public void destroy() {
